+++ NEWS.md
@@ -1,3 +1,11 @@
+### 0.1.0
+
+* Only supports Julia 0.4.x
+* Adds exogenous inputs in state space model transition and observation equations
+* Allows all system matrices (transition, control input, transition noise covariance,
+observation, feed-forward, observation noise covariance) to be time-dependent
+* Implements Durbin-Koopman Kalman smoothing algorithm
+
 ### 0.0.3
 
 * support floor of Julia 0.3 and non-inclusive ceiling of Julia 0.5-